+++ src/Shiny.Locations/CompatShims.cs
@@ -12,12 +12,12 @@ namespace Shiny
         static IGpsManager Current { get; } = ShinyHost.Resolve<IGpsManager>();
 
         public static bool IsListening => Current.IsListening;
-        public static AccessState GetCurrentStatus(bool background) => Current.GetCurrentStatus(background);
+        public static AccessState GetCurrentStatus(GpsRequest request) => Current.GetCurrentStatus(request);
         public static IObservable<IGpsReading?> GetLastReading() => Current.GetLastReading();
-        public static Task<AccessState> RequestAccess(bool backgroundMode) => Current.RequestAccess(backgroundMode);
+        public static Task<AccessState> RequestAccess(GpsRequest request) => Current.RequestAccess(request);
         public static Task StartListener(GpsRequest request) => Current.StartListener(request);
         public static Task StopListener() => Current.StopListener();
-        public static IObservable<AccessState> WhenAccessStatusChanged(bool forBackground) => Current.WhenAccessStatusChanged(forBackground);
+        public static IObservable<AccessState> WhenAccessStatusChanged(GpsRequest request) => Current.WhenAccessStatusChanged(request);
         public static IObservable<IGpsReading> WhenReading() => Current.WhenReading();
     }