+++ ChangeLog.md
@@ -2,6 +2,9 @@
 
 ## 1.0.1
 [Enhancement][Breaking] New way to set notification sounds which allows you to use system sounds - Notification.CustomSoundFilePath has been removed
+[Enhancement][Android] If you target android 10, the following now use full AndroidX support
+    * Shiny.Core - WorkManager replaces JobService under the hood
+    * Shiny.Notifications - Uses NotificationCompat
 
 ## 1.0.0
 Initial Release for